refactoring ee cli to use providers
