+++ ee/cli/cmd/digger/default.go
@@ -2,9 +2,10 @@ package main
 
 import (
 	"fmt"
-	"github.com/diggerhq/digger/cli/pkg/azure"
 	"github.com/diggerhq/digger/cli/pkg/digger"
+	"github.com/diggerhq/digger/cli/pkg/github"
 	"github.com/diggerhq/digger/cli/pkg/usage"
+	"github.com/diggerhq/digger/ee/cli/pkg/comment_updater"
 	"github.com/spf13/cobra"
 	"log"
 	"os"
@@ -20,22 +21,7 @@ var defaultCmd = &cobra.Command{
 		switch ci {
 		case digger.GitHub:
 			logLeader = os.Getenv("GITHUB_ACTOR")
-			gitHubCI(lock, PolicyChecker, BackendApi, ReportStrategy)
-		case digger.GitLab:
-			logLeader = os.Getenv("CI_PROJECT_NAME")
-			gitLabCI(lock, PolicyChecker, BackendApi, ReportStrategy)
-		case digger.Azure:
-			// This should be refactored in the future because in this way the parsing
-			// is done twice, both here and inside azureCI, a better solution might be
-			// to encapsulate it into a method on the azure package and then grab the
-			// value here and pass it into the azureCI call.
-			azureContext := os.Getenv("AZURE_CONTEXT")
-			parsedAzureContext, _ := azure.GetAzureReposContext(azureContext)
-			logLeader = parsedAzureContext.BaseUrl
-			azureCI(lock, PolicyChecker, BackendApi, ReportStrategy)
-		case digger.BitBucket:
-			logLeader = os.Getenv("BITBUCKET_STEP_TRIGGERER_UUID")
-			bitbucketCI(lock, PolicyChecker, BackendApi, ReportStrategy)
+			github.GitHubCI(lock, PolicyChecker, BackendApi, ReportStrategy, comment_updater.CommentUpdaterProviderAdvanced{})
 		case digger.None:
 			print("No CI detected.")
 			os.Exit(10)
